requirements.txt
streamlit
pandas
plotly
google-genai
python-dotenv
openpyxl
